requirements.txt
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.13.0
tqdm==4.66.1
opencv-python==4.7.0.72
streamlit==1.27.2

# Additional features
gtts==2.2.4
playsound==1.2.2
googletrans==4.0.0rc1
pytorch-grad-cam==1.4.0
pillow==9.5.0